# models/appointment.go
package models

import (
	"time"

	"gorm.io/gorm"
)

// Appointment represents a booking by a patient
//
// swagger:model
type Appointment struct {
	// ID is auto-generated by the database
	//
	// required: false
	ID uint `json:"id,omitempty" swaggerignore:"true"`

	// PatientID links the appointment to a patient
	//
	// required: true
	PatientID uint `json:"patient_id" example:"1" binding:"required"`

	// DateTime is the appointment time in RFC3339 format
	//
	// required: true
	DateTime time.Time `json:"date_time" swaggertype:"string" example:"2025-06-01T15:00:00Z" binding:"required"`

	// Reason for the appointment
	//
	// required: true
	Reason string `json:"reason" example:"Routine checkup" binding:"required"`

	CreatedAt time.Time      `json:"-"`
	UpdatedAt time.Time      `json:"-"`
	DeletedAt gorm.DeletedAt `gorm:"index" json:"-"`
}