# models/patient.go
package models

import (
	"time"

	"gorm.io/gorm"
)

// Patient represents a user in the healthcare system
//
// swagger:model
type Patient struct {
	// ID is auto-generated by the database
	ID uint `json:"id,omitempty" swaggerignore:"true"`

	// Name of the patient
	Name string `json:"name" binding:"required,min=2" example:"Erza Shala"`

	// Email of the patient
	Email string `json:"email" binding:"required,email" gorm:"unique" example:"erza@example.com"`

	// Appointments linked to the patient
	Appointments []Appointment `json:"appointments,omitempty" gorm:"foreignKey:PatientID" swaggerignore:"true"`

	CreatedAt time.Time      `json:"-"`
	UpdatedAt time.Time      `json:"-"`
	DeletedAt gorm.DeletedAt `gorm:"index" json:"-"`
}